+++ next.config.js
@@ -85,6 +85,9 @@ export default withLess(
 
       return config
     },
+    /**
+     * Comment this out if you're working on OG images.
+     */
     output: "export",
     images: {
       loader: "custom",
