+++ src/app/conf/2025/components/og-images/normalize-protocol-relative-url.tsx
@@ -1,0 +1,6 @@
+export function normalizeProtocolRelativeUrl(url: string) {
+  if (url.startsWith("//")) {
+    return `https:${url}`
+  }
+  return url
+}
