--- src/app/conf/2025/components/speaker-opengraph-image.tsx
@@ -1,374 +1,0 @@
-import type { SchedSpeaker } from "@/app/conf/2023/types"
-import { CalendarIcon } from "../../_design-system/pixelarticons/calendar-icon"
-import { PinIcon } from "../../_design-system/pixelarticons/pin-icon"
-import { formatSpeakerPosition } from "./format-speaker-position"
-import { GraphQLLogo } from "./graphql-conf-logo-link"
-
-export const colors = {
-  neu0: "hsl(0 0% 100%)",
-  neu50: "hsl(75 57% 97%)",
-  neu100: "hsl(75 15% 95%)",
-  neu200: "hsl(77 14% 90%)",
-  neu300: "hsl(76 14% 85%)",
-  neu400: "hsl(77 14% 80%)",
-  neu500: "hsl(74 14% 70%)",
-  neu600: "hsl(76 15% 60%)",
-  neu700: "hsl(76 15% 40%)",
-  neu800: "hsl(77 14% 20%)",
-  neu900: "hsl(75 15% 5%)",
-  secLighter: "hsl(79 80% 90%)",
-  priBase: "hsl(319 100% 44.1%)",
-}
-
-export const fonts = {
-  sans: "'Host Grotesk', var(--font-sans)",
-  mono: "'Commit Mono', var(--font-mono)",
-}
-
-const RIGHT_COLUMN_WIDTH_PX = 476
-
-export interface SpeakerOpengraphImageProps
-  extends React.HTMLAttributes<HTMLElement> {
-  speaker: SchedSpeaker
-  date: string
-  year: string
-  location: string
-}
-
-export function SpeakerOpengraphImage({
-  speaker,
-  date,
-  year,
-  location,
-  ...rest
-}: SpeakerOpengraphImageProps) {
-  return (
-    <article
-      style={{
-        display: "flex",
-        height: "630px",
-        width: "1200px",
-        flexDirection: "column",
-        overflow: "hidden",
-        borderWidth: "2px",
-        backgroundColor: colors.neu100,
-        borderColor: colors.neu600,
-        fontFamily: fonts.sans,
-      }}
-      {...rest}
-    >
-      <ConferenceOpengraphImageHeader
-        year={year}
-        date={date}
-        location={location}
-      />
-
-      <div style={{ display: "flex" }}>
-        <div
-          style={{
-            display: "flex",
-            flex: 1,
-            flexDirection: "column",
-            justifyContent: "center",
-          }}
-        >
-          <div
-            style={{
-              display: "flex",
-              flex: 1,
-              flexDirection: "column",
-              justifyContent: "center",
-              gap: "2.5rem",
-              padding: "2.5rem",
-              paddingRight: "4rem",
-            }}
-          >
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-                gap: "2.5rem",
-              }}
-            >
-              <h3
-                style={{
-                  margin: 0,
-                  fontSize: "88px",
-                  fontWeight: "normal",
-                  lineHeight: "1.25",
-                  color: colors.neu900,
-                }}
-              >
-                {speaker.name}
-              </h3>
-
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  gap: "2rem",
-                }}
-              >
-                <span
-                  style={{
-                    fontFamily: fonts.sans,
-                    fontSize: "32px",
-                    fontWeight: "normal",
-                    lineHeight: "1.25",
-                    color: colors.neu700,
-                  }}
-                >
-                  {formatSpeakerPosition(speaker)}
-                </span>
-              </div>
-            </div>
-          </div>
-
-          <OpengraphImageFooter>Speakers</OpengraphImageFooter>
-        </div>
-
-        {speaker.avatar && (
-          <div
-            style={{
-              position: "relative",
-              display: "flex",
-              overflow: "hidden",
-              borderLeft: `2px solid ${colors.neu600}`,
-              width: RIGHT_COLUMN_WIDTH_PX + 2,
-              height: RIGHT_COLUMN_WIDTH_PX,
-            }}
-          >
-            <img
-              src={normalizeProtocolRelativeUrl(speaker.avatar)}
-              alt=""
-              style={{
-                objectFit: "cover",
-                // @vercel/og doesn't support mix blend mode
-                // and SVG `<image>` is crashing similarly to
-                // https://github.com/vercel/satori/issues/650.
-                // So we use `sepia` and `hue-rotate` to change
-                // the hue to around 79deg from the design system's
-                // secondary color (yellowish green).
-                filter: "sepia(1) hue-rotate(37.5deg)",
-              }}
-              width={RIGHT_COLUMN_WIDTH_PX}
-              height={RIGHT_COLUMN_WIDTH_PX}
-            />
-            <div
-              style={{
-                position: "absolute",
-                top: 0,
-                left: 0,
-                width: RIGHT_COLUMN_WIDTH_PX,
-                height: RIGHT_COLUMN_WIDTH_PX,
-                backgroundColor: colors.secLighter,
-                opacity: 0.25,
-              }}
-            />
-          </div>
-        )}
-      </div>
-    </article>
-  )
-}
-
-export function ConferenceOpengraphImageHeader({
-  year,
-  date,
-  location,
-}: {
-  year: string
-  date: string
-  location: string
-}) {
-  return (
-    <header
-      style={{
-        display: "flex",
-        alignItems: "center",
-        borderBottom: `2px solid ${colors.neu600}`,
-      }}
-    >
-      <div
-        style={{
-          display: "flex",
-          flex: 1,
-          alignItems: "center",
-          gap: "1.5rem",
-          borderRight: `2px solid ${colors.neu600}`,
-          padding: "2.5rem",
-          paddingRight: "4rem",
-          height: "154px",
-        }}
-      >
-        <div style={{ display: "flex", alignItems: "center", gap: "1rem" }}>
-          <div
-            style={{
-              fontFamily: fonts.mono,
-              display: "flex",
-              fontWeight: "normal",
-              textTransform: "uppercase",
-              lineHeight: "1",
-              color: colors.neu900,
-            }}
-          >
-            <div
-              style={{
-                display: "flex",
-                height: "74px",
-                alignItems: "center",
-                gap: "1rem",
-                fontSize: "40px",
-                lineHeight: "1",
-                textTransform: "uppercase",
-              }}
-            >
-              <GraphQLLogo
-                style={{
-                  height: "3rem",
-                  width: "3rem",
-                  color: colors.priBase,
-                  /* hack: satori aligns this SVG differently than browsers, it will look off center in /workroom,
-                           but centered in the images */
-                  marginTop: "-6px",
-                }}
-              />
-              <span>/</span>
-              <div
-                style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}
-              >
-                GraphQLConf{" "}
-                <span style={{ color: colors.priBase }}>{year}</span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div
-        style={{
-          display: "flex",
-          height: "100%",
-          flexShrink: 0,
-          flexDirection: "column",
-          justifyContent: "center",
-          width: RIGHT_COLUMN_WIDTH_PX,
-        }}
-      >
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            gap: "1.5rem",
-            borderBottom: `2px solid ${colors.neu600}`,
-            paddingLeft: "1.5rem",
-            paddingRight: "1.5rem",
-            paddingTop: "26px",
-            paddingBottom: "26px",
-          }}
-        >
-          <div style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}>
-            <CalendarIcon
-              width="24"
-              height="24"
-              style={{
-                /* hack: different across satori and browsers */
-                transform: "translateY(-3px)",
-                color: colors.priBase,
-              }}
-            />
-            <span
-              style={{
-                fontFamily: fonts.mono,
-                display: "flex",
-                fontSize: "1.25rem",
-                fontWeight: "normal",
-                textTransform: "uppercase",
-                lineHeight: "1.2",
-                color: colors.neu900,
-              }}
-            >
-              {date}, {year}
-            </span>
-          </div>
-        </div>
-
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            gap: "1.5rem",
-            paddingLeft: "1.5rem",
-            paddingRight: "1.5rem",
-            paddingTop: "26px",
-            paddingBottom: "26px",
-          }}
-        >
-          <div style={{ display: "flex", alignItems: "center", gap: "0.5rem" }}>
-            <PinIcon
-              width="24"
-              height="24"
-              style={{
-                /* hack: different across satori and browsers */
-                transform: "translateY(-2px)",
-                color: colors.priBase,
-              }}
-            />
-            <span
-              style={{
-                fontFamily: fonts.mono,
-                fontSize: "1.25rem",
-                fontWeight: "normal",
-                textTransform: "uppercase",
-                lineHeight: "1.2",
-                color: colors.neu900,
-              }}
-            >
-              {location}
-            </span>
-          </div>
-        </div>
-      </div>
-    </header>
-  )
-}
-
-export function normalizeProtocolRelativeUrl(url: string) {
-  if (url.startsWith("//")) {
-    return `https:${url}`
-  }
-  return url
-}
-
-export function OpengraphImageFooter({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <footer
-      style={{
-        display: "flex",
-        alignItems: "center",
-        borderTop: `2px solid ${colors.neu600}`,
-        paddingLeft: "2.5rem",
-        paddingRight: "4rem",
-        paddingTop: "2rem",
-        paddingBottom: "2rem",
-      }}
-    >
-      <span
-        style={{
-          fontFamily: fonts.mono,
-          fontSize: "1.5rem",
-          fontWeight: "normal",
-          textTransform: "uppercase",
-          lineHeight: "1",
-          color: colors.neu900,
-        }}
-      >
-        {children}
-      </span>
-    </footer>
-  )
-}
