+++ src/app/conf/2025/opengraph-image.tsx
@@ -1,0 +1,10 @@
+import { SimpleOpengraphImage } from "./components/og-images/simple-opengraph-image"
+export {
+  generateStaticParams,
+  contentType,
+  size,
+} from "./components/og-images/simple-opengraph-image"
+
+export default SimpleOpengraphImage.bind(null, {
+  pageTitle: "GraphQLConf 2025",
+})
